requirements.txt
Flask>=2.0
Flask-Uploads>=0.2.1
PyPDF2>=3.0.0
python-docx>=0.8.11

# Text processing & ML
transformers
sentence-transformers>=2.2.2
scikit-learn>=1.2.0
datasets>=2.20.0
peft>=0.11.1
accelerate==0.34.2

# Web scraping / Job board access
requests>=2.28.0
beautifulsoup4>=4.11.1
lxml>=4.9.2

# Utilities
python-dotenv>=1.0.0
gunicorn>=20.1.0